fix(lottery): Use separate node calls to fetch ticket data for one round
