+++ src/state/lottery/getUserLotteryData.ts
@@ -2,8 +2,9 @@ import { request, gql } from 'graphql-request'
 import { GRAPH_API_LOTTERY } from 'config/constants/endpoints'
 import { LotteryTicket } from 'config/constants/types'
 import { LotteryUserGraphEntity, LotteryResponse, UserRound } from 'state/types'
-import { getRoundIdsArray, fetchMultipleLotteries, hasRoundBeenClaimed, processRawTicketsResponse } from './helpers'
+import { getRoundIdsArray, fetchMultipleLotteries, hasRoundBeenClaimed } from './helpers'
 import { fetchUserTicketsForMultipleRounds } from './fetchUnclaimedUserRewards'
+import { processRawTicketsResponse } from './getUserInfoForLotteryId'
 
 const applyNodeDataToUserGraphResponse = (
   nodeData: LotteryResponse[],
