+++ src/views/Lottery/components/ViewTicketsModal/PreviousRoundTicketsInner.tsx
@@ -15,8 +15,9 @@ import {
 import styled from 'styled-components'
 import { useWeb3React } from '@web3-react/core'
 import { LotteryTicket, LotteryTicketClaimData } from 'config/constants/types'
-import { fetchLottery, fetchTickets } from 'state/lottery/helpers'
+import { fetchLottery } from 'state/lottery/helpers'
 import { getWinningTickets } from 'state/lottery/fetchUnclaimedUserRewards'
+import { getUserInfoForLotteryId } from 'state/lottery/getUserInfoForLotteryId'
 import { LotteryRound } from 'state/types'
 import { useGetUserLotteryGraphRoundById } from 'state/hooks'
 import { useTranslation } from 'contexts/Localization'
@@ -108,7 +109,7 @@ const PreviousRoundTicketsInner: React.FC<{ roundId: string }> = ({ roundId }) =
     }
 
     const fetchData = async () => {
-      const userTickets = await fetchTickets(account, roundId, totalTickets)
+      const userTickets = await getUserInfoForLotteryId(account, roundId)
       const lotteryData = await fetchLottery(roundId)
       const processedLotteryData = processLotteryResponse(lotteryData)
       const winningTickets = await getWinningTickets({
